Create end of message protocol fix
Now the message "---OK---" is sent to mark the end of a transmission.

This may have to be revised to take into account the possibility of a
file having "---OK---" at the end of it. The result would be that it
would be omitted from the transferred file.

There is also a chance that "---OK---" is found somewhere in the middle
of a message and this would cause a premature termination.