Create successful binary protocol
Now the protocol is as follows:

Data is sent in chunks of 16bytes.

This 16byte chunk is made up of: (data | length | flag)
Where data=14bytes
    length=1byte
      flag=1byte

This allows the receiver to check the flag to check if last transmission
(i.e. a sort of "over" as in radio comms).

The reason a length < 14 can not be used as an end of tranmission
signifier is in the cases where the data to be sent can be divided by 14
exactly. Then the last message would still be 14bytes, and the receiver
would still wait for more data.