Add test encryption and refactor file sizes
Test encryption was successful (using hardcoded IV and Key, and with no
socket data transfer)

However, because our plaintext data size is 16bytes the cipher must add
another block of padding to indicate length. This is inefficient (i.e.
transfering n data chunks each of size chunk*2)

To change this data chunk size is not simple as the code is messy and
hard coded. So to fix this, refactoring has started that uses a header
file with TOTAL_SIZE constants. The idea is that once everything uses
these constants, it will be easy to experiment with data block size
changes, and both client and server can use a copy of the same header
file.

Currently the server supports this, but the client does not.