Fix file upload name bug
The issue was that after refactoring the code in processor.h, the
filename extraction was happening after the length variable was changed.

Additionally, the 434 --> BLOCK_SIZE change bug was caused by the large
434 value covering up a typo.

The sent data will always be of size "chunk_size" even if 0 bytes are
read. This is to allow the flags to still be used. Instead, the
encrypted char[] was being assigned a value of "read+BLOCK_SIZE", when
it should be "chunk_size+BLOCK_SIZE".