Fix slow speed issue
The speed bottleneck was the overhead caused by opening/closing the file
for every write.

The buffer sizes were also increased.

However, it appears there is stil a residual bug that sometimes causes
encryption errors. This is most likely due to the buffer size changes in
constants.h

To do:

 - refactor the write_file function in file.h
 - ensure a similar bottleneck does not occur for reading files from the
server
 - ensure changing buffer sizes are not causing encryption bugs
 - investigate the lack of write errors when disk quota is exceeded
 - give program final test
 - create write up document with timing tests