Creat test script and fix bad key error
The bad key error was caused by the socket not always having the correct
number of bytes to read (i.e. the full encrypted block)

The fix was to detect this case by checking the read length response,
and if under the required size, a loop was entered to continue reading
from the socket until the buffer is filled up.

Additionally, a basic test script was created that helped find this and
will be used eventually for creating an average of all different command
options using awk/grep/cut etc.

Ideally it will also compare hashes to ensure each test passes.