Finish test script, collect results, fix null bug
The test script has finished. It needs to be run separately for each
cipher in order to collect the different results in different text
files.

The results for each case have been collected in the results directory.
These still need to be processed but a python script and bash cut script
have been created which will be able to generate results easily.

Lastly, when checking the programs functionality it was noticed that the
null cipher option was not working properly. This was due to the null
cipher settings sending chunks of TOTAL_SIZE, not ENCRYPTED_SIZE.
Through normal encryption, this extra padding is added. To fix the bug,
this padding had to be "recreated" in the encryption.cpp file.

To do:
 - parse test results and extract medians for each case